Modify DateTimeAxis example
I thought the example for DateTimeAxis was too sparse, it was difficult to understand where/how to specify the start/end date and how to format the labels. I also fixed a typo where a LinearAxis was added to the PlotModel, not a DateTimeAxis. 

This example should make it easier for users to visualize how to get from two DateTimes to the range of labels that they want on the axis they provide.